fix(frontend): preserve braces during enrichment streaming
An ambiguous two-brace suffix could consume a Java closing brace while
assistant content was still streaming. Make enrichment closure state-aware
and isolate assistant markdown rendering from message chrome.

- use separate complete and streaming Marked instances
- cover partial delimiters and literal-brace completion
- extract assistant rendering below the component size limit
