+++ frontend/src/lib/components/MessageBubble.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
+  import AssistantMarkdownBody from './AssistantMarkdownBody.svelte'
   import type { ChatMessage } from '../services/chat'
   import { isRecoverableCsrfErrorMessage } from '../services/csrf'
-  import { parseMarkdown, applyJavaLanguageDetection } from '../services/markdown'
-  import { createDebouncedHighlighter } from '../utils/highlight'
 
   interface Props {
     message: ChatMessage
@@ -12,40 +11,14 @@
 
   let { message, index, isStreaming = false }: Props = $props()
 
-  let contentEl: HTMLElement | null = $state(null)
-
   /** Visual feedback state for clipboard operations. */
   let copyState = $state<'idle' | 'success' | 'error'>('idle')
-
-  // Debounced highlighter with automatic cleanup
-  const { scheduleHighlight, cleanup: cleanupHighlighter } = createDebouncedHighlighter()
-
-  // Render markdown synchronously - SSR-safe parsing without DOM operations
-  let renderedContent = $derived(
-    message.role === 'assistant' && message.messageText
-      ? parseMarkdown(message.messageText)
-      : ''
-  )
 
   let showCsrfRefreshButton = $derived(
     message.role === 'assistant' &&
       message.isError === true &&
       isRecoverableCsrfErrorMessage(message.messageText)
   )
-
-  // Apply Java language detection and highlight code blocks after render
-  // Debounced to avoid flicker during streaming
-  $effect(() => {
-    if (renderedContent && contentEl) {
-      if (isStreaming) {
-        return cleanupHighlighter
-      }
-      // Apply Java language detection before highlighting (client-side DOM operation)
-      applyJavaLanguageDetection(contentEl)
-      scheduleHighlight(contentEl, isStreaming)
-    }
-    return cleanupHighlighter
-  })
 
   /** Duration to show copy feedback before returning to idle. */
   const COPY_FEEDBACK_DURATION_MS = 1500
@@ -97,14 +70,7 @@
     {#if message.role === 'user'}
       <p class="user-text">{message.messageText}</p>
     {:else}
-      <div class="assistant-content" bind:this={contentEl}>
-        {#if renderedContent}
-          {@html renderedContent}
-        {:else}
-          <p>{message.messageText}</p>
-        {/if}
-        <span class="cursor" class:visible={isStreaming}></span>
-      </div>
+      <AssistantMarkdownBody markdown={message.messageText} {isStreaming} />
       {#if showCsrfRefreshButton}
         <button type="button" class="csrf-refresh-btn" onclick={reloadCurrentPage}>
           Refresh and retry
@@ -205,9 +171,8 @@
     color: var(--color-error);
   }
 
-  /* Shared text wrapping for long unbroken strings (URLs, code identifiers) */
-  .user-text,
-  .assistant-content {
+  /* Wrap long unbroken user strings such as URLs. */
+  .user-text {
     overflow-wrap: break-word;
     word-break: break-word;
   }
@@ -217,128 +182,6 @@
     font-size: var(--text-base);
     line-height: var(--leading-relaxed);
     margin: 0;
-  }
-
-  /* Assistant content - prose styles */
-  .assistant-content {
-    font-size: var(--text-base);
-    line-height: var(--leading-relaxed);
-    color: var(--color-text-primary);
-  }
-
-  .assistant-content :global(p) {
-    margin: 0 0 var(--space-4);
-  }
-
-  .assistant-content :global(p:last-child) {
-    margin-bottom: 0;
-  }
-
-  .assistant-content :global(h1),
-  .assistant-content :global(h2),
-  .assistant-content :global(h3),
-  .assistant-content :global(h4) {
-    font-family: var(--font-serif);
-    font-weight: 500;
-    margin: var(--space-6) 0 var(--space-3);
-    letter-spacing: var(--tracking-tight);
-  }
-
-  .assistant-content :global(h1:first-child),
-  .assistant-content :global(h2:first-child),
-  .assistant-content :global(h3:first-child) {
-    margin-top: 0;
-  }
-
-  .assistant-content :global(h2) {
-    font-size: var(--text-xl);
-  }
-
-  .assistant-content :global(h1) {
-    font-size: var(--text-2xl);
-  }
-
-  .assistant-content :global(h3) {
-    font-size: var(--text-lg);
-  }
-
-  .assistant-content :global(h4) {
-    font-size: var(--text-base);
-  }
-
-  .assistant-content :global(ul),
-  .assistant-content :global(ol) {
-    margin: 0 0 var(--space-4);
-    padding-left: var(--space-6);
-  }
-
-  .assistant-content :global(li) {
-    margin-bottom: var(--space-2);
-  }
-
-  .assistant-content :global(li:last-child) {
-    margin-bottom: 0;
-  }
-
-  .assistant-content :global(strong) {
-    font-weight: 600;
-    color: var(--color-text-primary);
-  }
-
-  .assistant-content :global(a) {
-    color: var(--color-accent);
-    text-decoration: underline;
-    text-underline-offset: 2px;
-  }
-
-  .assistant-content :global(a:hover) {
-    text-decoration-thickness: 2px;
-  }
-
-  .assistant-content :global(blockquote) {
-    margin: var(--space-4) 0;
-    padding: var(--space-3) var(--space-4);
-    border-left: 3px solid var(--color-accent);
-    background: var(--color-surface-subtle);
-    border-radius: 0 var(--radius-md) var(--radius-md) 0;
-    font-style: italic;
-    color: var(--color-text-secondary);
-  }
-
-  /* Code blocks in assistant messages */
-  .assistant-content :global(pre) {
-    margin: var(--space-4) 0;
-    padding: var(--space-4);
-    background: var(--color-bg-primary);
-    border: 1px solid var(--color-border-subtle);
-    border-radius: var(--radius-lg);
-    overflow-x: auto;
-    font-size: var(--text-sm);
-  }
-
-  .assistant-content :global(pre code) {
-    background: none;
-    padding: 0;
-    border: none;
-    font-size: inherit;
-    color: var(--color-text-primary);
-  }
-
-  /* Streaming cursor - always present, visibility controlled by class */
-  .cursor {
-    display: inline-block;
-    width: 2px;
-    height: 1.2em;
-    background: var(--color-accent);
-    margin-left: 2px;
-    vertical-align: text-bottom;
-    opacity: 0;
-    transition: opacity 150ms ease-out;
-  }
-
-  .cursor.visible {
-    opacity: 1;
-    animation: typing-cursor 0.8s ease-in-out infinite;
   }
 
   .csrf-refresh-btn {
@@ -359,11 +202,6 @@
 
   .csrf-refresh-btn:hover {
     background: var(--color-bg-hover);
-  }
-
-  @keyframes typing-cursor {
-    0%, 100% { opacity: 1; }
-    50% { opacity: 0; }
   }
 
   /* Actions - desktop hover only (never on touch / narrow viewports) */
@@ -467,23 +305,6 @@
     .avatar svg {
       width: 16px;
       height: 16px;
-    }
-
-    .assistant-content {
-      font-size: var(--text-sm);
-    }
-
-    .assistant-content :global(pre) {
-      padding: var(--space-3);
-      font-size: var(--text-xs);
-      margin: var(--space-3) calc(-1 * var(--space-3));
-      border-radius: 0;
-      border-left: none;
-      border-right: none;
-    }
-
-    .assistant-content :global(code:not(pre code)) {
-      font-size: 0.8em;
     }
   }
